feat: изменение сущности заказа и логики создания

+++ ProductService/src/main/java/com/example/productservice/dto/ProductDto.java
@@ -4,6 +4,7 @@ import jakarta.persistence.Column;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 @Data
 @AllArgsConstructor
@@ -11,7 +12,7 @@ public class ProductDto {
     private Long id;
     private String name;
     private String description;
-    private Long price;
+    private BigDecimal price;
     private Integer stockQuantity;
     private LocalDate createdAt;
